Rename types to _types
