+++ src/__configuration__/themes/women-s-day.ts
@@ -3,7 +3,7 @@ import * as Colors from '@brightlayer-ui/colors';
 import Banner from '../../app/assets/themes/women-s-day/international-women-s-day-banner.svg';
 import AppBarTile from '../../app/assets/themes/women-s-day/international-women-s-day-appbar-tile.svg';
 import Cursor from '../../app/assets/themes/women-s-day/women-s-day-cursor.png';
-import { Schedule } from './types';
+import { Schedule } from './_types';
 
 // Custom palette since we don't have pink colors
 const LIGHT_COLOR = '#F1C4DA';
